Adding module initialization on app initialization and on URL change.
In the older version, when the app initialized, all of the modules initialized too. Of course, we do not want it (because then in every page refresh, every HTTP request would be called, even when we do not show its output). In the corrected version the module's InitData message (if the module has one) is being called if the module is loaded (by page load or URL change).